requirements.txt
absl-py==1.0.0
aiofiles==0.6.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
anyio==3.4.0
apache-airflow==2.2.3
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-google==6.2.0
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-postgres==2.4.0
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.3
asgiref==3.5.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==20.3.0
Babel==2.9.1
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
blinker==1.4
blis==0.7.5
Brotli==1.0.9
brotli-asgi==1.1.0
cachetools==4.2.2
catalogue==2.0.6
cattrs==1.5.0
certifi==2020.12.5
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
command-not-found==0.3
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.8
cycler==0.11.0
cymem==2.0.6
datasets==1.18.3
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.1.0
docutils==0.16
ecdsa==0.14.1
email-validator==1.1.3
ez-setup==0.9
fastapi==0.63.0
filelock==3.4.0
Flask==1.1.2
Flask-AppBuilder==3.4.1
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flatbuffers==2.0
fonttools==4.31.1
frozenlist==1.3.0
fsspec==2022.1.0
gast==0.5.3
gensim==4.1.2
google-ads==14.0.0
google-api-core==1.31.4
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-automl==2.5.2
google-cloud-bigquery==2.31.0
google-cloud-bigquery-datatransfer==3.4.1
google-cloud-bigquery-storage==2.10.1
google-cloud-bigtable==1.7.0
google-cloud-build==3.7.1
google-cloud-container==1.0.1
google-cloud-core==1.7.2
google-cloud-datacatalog==3.6.1
google-cloud-dataproc==3.1.1
google-cloud-dataproc-metastore==1.3.1
google-cloud-dlp==1.0.0
google-cloud-kms==2.10.1
google-cloud-language==1.3.0
google-cloud-logging==2.7.0
google-cloud-memcache==1.0.0
google-cloud-monitoring==2.8.0
google-cloud-os-login==2.5.1
google-cloud-pubsub==2.9.0
google-cloud-redis==2.5.0
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.19.1
google-cloud-speech==1.3.2
google-cloud-storage==1.43.0
google-cloud-tasks==2.7.1
google-cloud-texttospeech==1.0.1
google-cloud-translate==1.7.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-cloud-workflows==1.5.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
graphviz==0.19.1
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-gcp==0.2.2
gunicorn==20.1.0
h11==0.9.0
h5py==3.6.0
httpcore==0.11.1
httplib2==0.19.1
httptools==0.3.0
httpx==0.15.5
huggingface-hub==0.4.0
hurry.filesize==0.9
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
inflection==0.5.1
ipython_genutils==0.2.0
iso8601==1.0.2
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==3.0.3
joblib==1.1.0
json-merge-patch==0.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-core==4.6.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
langcodes==3.3.0
language-selector==0.1
lazy-object-proxy==1.4.3
libclang==13.0.0
libcst==0.3.23
lockfile==0.12.2
loguru==0.6.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.1
more-itertools==4.2.0
multidict==6.0.2
multiprocess==0.70.12.2
murmurhash==1.0.6
mypy-extensions==0.4.3
nbformat==5.0.4
netifaces==0.10.4
nltk==3.7
nox==2020.12.31
numpy==1.20.3
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
opensearch-py==1.0.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandas-gbq==0.14.1
passlib==1.7.4
pathy==0.6.1
pendulum==2.1.2
pexpect==4.6.0
Pillow==9.0.1
platformdirs==2.4.0
preshed==3.0.6
prison==0.2.1
proto-plus==1.18.1
protobuf==3.19.1
psutil==5.8.0
psycopg2-binary==2.9.2
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.2.0/pt_core_news_sm-3.2.0-py3-none-any.whl
py==1.11.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pydata-google-auth==1.3.0
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-daemon==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python-jose==3.2.0
python-multipart==0.0.5
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
regex==2022.1.18
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==10.16.0
rsa==4.8
rubrix==0.12.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
setproctitle==1.2.2
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
sos==4.2
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.9
srsly==2.4.2
starlette==0.13.6
stopwordsiso==0.6.1
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.13
threadpoolctl==3.1.0
tokenizers==0.11.4
torch==1.10.2
tqdm==4.62.3
traitlets==4.3.3
transformers==4.16.2
typer==0.4.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
ubuntu-advantage-tools==27.5
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.7
uvicorn==0.17.4
uvloop==0.16.0
virtualenv==20.10.0
wasabi==0.9.0
watchgod==0.7
websockets==10.1
Werkzeug==1.0.1
wrapt==1.13.3
WTForms==2.3.3
xxhash==2.0.2
yarl==1.7.2
zipp==3.6.0
